organized app in relation to redux and added the chrome redux devtool extension
