+++ dragonStack/backend/package.json
@@ -18,6 +18,7 @@
 		"cors": "^2.8.5",
 		"event-stream": "^3.3.4",
 		"express": "^4.16.4",
+		"node": "^11.2.0",
 		"pg": "^7.6.1"
 	}
 }
