+++ dragonStack/frontend/src/actions/generation.js
@@ -1,0 +1,8 @@
+import { GENERATION_ACTION_TYPE } from "./types";
+
+export const generationActionCreator = payload => {
+  return {
+    type: GENERATION_ACTION_TYPE,
+    generation: payload
+  };
+};
