+++ dragonStack/frontend/src/actions/types.js
@@ -1,0 +1,1 @@
+export const GENERATION_ACTION_TYPE = "GENERATION_ACTION_TYPE";
