+++ dragonStack/frontend/src/index.js
@@ -3,49 +3,16 @@ import { createStore } from "redux";
 import { render } from "react-dom";
 import Generation from "./componets/Generation";
 import Dragon from "./componets/Dragon";
+import { generationReducer } from "./reducers";
+import { generationActionCreator } from "./actions/generation";
 import "./index.css";
 
-const DEFAULT_GENERATION = { generationId: "", expiration: "" };
-
-const GENERATION_ACTION_TYPE = "GENERATION_ACTION_TYPE";
-
-const generationReducer = (state, action) => {
-  if (action.type === GENERATION_ACTION_TYPE) {
-    return { generation: action.generation };
-  }
-
-  return {
-    generation: DEFAULT_GENERATION
-  };
-};
-
-const store = createStore(generationReducer);
+const store = createStore(
+  generationReducer,
+  window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__()
+);
 
 store.subscribe(() => console.log("store state update", store.getState()));
-
-console.log("store", store);
-
-store.dispatch({ type: "foo" });
-store.dispatch({
-  type: GENERATION_ACTION_TYPE,
-  generation: { generationId: "goo", expiration: "bar" }
-});
-
-console.log("store.getState()", store.getState());
-
-const generationActionCreator = payload => {
-  return {
-    type: GENERATION_ACTION_TYPE,
-    generation: payload
-  };
-};
-
-const zooAction = generationActionCreator({
-  generationId: "zoo",
-  expiration: "bar"
-});
-
-store.dispatch(zooAction);
 
 fetch("http://localhost:3000/generation")
   .then(response => response.json())
@@ -55,7 +22,7 @@ fetch("http://localhost:3000/generation")
 
 render(
   <div>
-    <h2>Dragon Stack</h2>
+    <h2>Dragon Stack from React</h2>
     <Generation />
     <Dragon />
   </div>,
